refactor: rename and fix margins
